Reorganize providers tests
