+++ ckanext/search/tests/conftest.py
@@ -2,6 +2,8 @@ import pytest
 
 from ckan.plugins.toolkit import config
 
+from ckanext.search.index import clear_index
+
 
 def pytest_addoption(parser):
     parser.addoption(
@@ -15,6 +17,7 @@ def pytest_addoption(parser):
 def search_providers():
     pass
 
+
 def pytest_runtest_setup(item):
 
     if "search_providers" in item.fixturenames and item.config.option.search_provider:
@@ -23,3 +26,10 @@ def pytest_runtest_setup(item):
                 "ckan.search.search_backend", item.config.option.search_provider
             )
         )
+
+
+@pytest.fixture
+def clean_search_index():
+    clear_index()
+    yield
+    clear_index()
